optimize playback loop to eliminate busy waiting
dizi used to poll for input every X interval. This is to keep
the playback loop going so the song can play. But if a user
pauses the playback loop, we should not continuously ask for
input on regular intervals. Instead, we should poll for input
and do nothing when there is no input. Only when we receive
input for resuming playback should we change back to regular
interval input polling